Move partition - record serialisation into StateStoreFileUtils
